remove rank-1 scaling
